Abstract base section for assessments
Both risk and healthcare sections behave the same way. This work
abstracts tha behaviour into a single base section class.
